backend/requirements.txt
# backend/requirements.txt (Mise à jour)
fastapi
uvicorn[standard]
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
sqlalchemy
asyncpg
python-multipart
boto3
pytesseract
Pillow
tenacity
email-validator